+++ src/user-data/jsonFileStore.js
@@ -1,8 +1,8 @@
 import { get, set } from 'lodash-es';
-import { getConfigValue } from './config.js';
-import { logger } from './logger.js';
-import { loadFileContents, saveFile } from './io.js';
-import { DataFileNotFoundError } from './errors/index.js';
+import { getConfigValue } from '../config.js';
+import { logger } from '../logger.js';
+import { loadFileContents, saveFile } from '../io.js';
+import { DataFileNotFoundError } from '../errors/index.js';
 
 const dataFilePath = getConfigValue('paths.dataStoreFile');
